[BUILD_RUN] added options -skipbuild and -skiprun

+++ build_run.sh
@@ -14,6 +14,8 @@
 BUILD_OPT="nofpga"
 # BUILD_OPT="fpga"
 NO_NINJA=false
+DO_BUILD=true
+DO_RUN=true
 
 # Check if a command-line argument is passed and override BUILD_OPT
 if [ "$#" -eq 0 ]; then
@@ -23,6 +25,12 @@ else
         case $arg in
             "fpga"|"nofpga")
                 BUILD_OPT="$arg"
+                ;;
+            "-skipbuild")
+                DO_BUILD=false
+                ;;
+            "-skiprun")
+                DO_RUN=false
                 ;;
             "-noninja")
                 NO_NINJA=true
@@ -35,33 +43,37 @@ else
     done
 fi
 
-rm -rf build_${BUILD_OPT}
+if [ "$DO_BUILD" = true ]; then
+    rm -rf build_${BUILD_OPT}
 
-if [ "$NO_NINJA" = true ]; then
-    USE_FPGA="OFF"
-    if [ "$BUILD_OPT" = "fpga" ]; then
-        USE_FPGA="ON"
+    if [ "$NO_NINJA" = true ]; then
+        USE_FPGA="OFF"
+        if [ "$BUILD_OPT" = "fpga" ]; then
+            USE_FPGA="ON"
+        fi
+        
+        mkdir build_${BUILD_OPT}
+        cd build_${BUILD_OPT}
+        cmake .. -DCMAKE_C_FLAGS="-mno-avx2 -mno-avx -mno-fma -mno-sse3 -mno-ssse3" \
+                    -DCMAKE_CXX_FLAGS="-mno-avx2 -mno-avx -mno-fma -mno-sse3 -mno-ssse3" \
+                    -DGGML_METAL=OFF -DGGML_BLAS=OFF -DGGML_LLAMAFILE=OFF \
+                    -DGGML_COMPILER_SUPPORT_MATMUL_INT8=OFF -DGGML_COMPILER_SUPPORT_FP16_VECTOR_ARITHMETIC=OFF \
+                    -DGGML_COMPILER_SUPPORT_DOTPROD=OFF -DGGML_NATIVE=OFF -DGGML_ACCELERATE=OFF -DGGML_SIMD=OFF \
+                    -DFPGA_ACCELERATOR=$USE_FPGA
+        make
+        cd ..
+    else
+        cmake --preset ${BUILD_OPT}
+        cmake --build --preset ${BUILD_OPT}
     fi
-    
-    mkdir build_${BUILD_OPT}
-    cd build_${BUILD_OPT}
-    cmake .. -DCMAKE_C_FLAGS="-mno-avx2 -mno-avx -mno-fma -mno-sse3 -mno-ssse3" \
-                -DCMAKE_CXX_FLAGS="-mno-avx2 -mno-avx -mno-fma -mno-sse3 -mno-ssse3" \
-                -DGGML_METAL=OFF -DGGML_BLAS=OFF -DGGML_LLAMAFILE=OFF \
-                -DGGML_COMPILER_SUPPORT_MATMUL_INT8=OFF -DGGML_COMPILER_SUPPORT_FP16_VECTOR_ARITHMETIC=OFF \
-                -DGGML_COMPILER_SUPPORT_DOTPROD=OFF -DGGML_NATIVE=OFF -DGGML_ACCELERATE=OFF -DGGML_SIMD=OFF \
-                -DFPGA_ACCELERATOR=$USE_FPGA
-    make
-    cd ..
-else
-    cmake --preset ${BUILD_OPT}
-    cmake --build --preset ${BUILD_OPT}
-fi
+fi # dobuild
 
-MODEL="./dir/to/model.gguf"
-SEED=315
-# If dumping out dotp info (default off), then use NUMGEN=1
-NUMGEN=10
-MODEL_PROMPT="Once upon a time there was a fox who "
+if [ "$DO_RUN" = true ]; then
+    MODEL="./dir/to/model.gguf"
+    SEED=315
+    # If dumping out dotp info (default off), then use NUMGEN=1
+    NUMGEN=10
+    MODEL_PROMPT="Once upon a time there was a fox who "
 
-./build_${BUILD_OPT}/bin/llama-cli -t 1 -m "$MODEL" -s "$SEED" -n "$NUMGEN" -e -p "$MODEL_PROMPT" # running with 1 thread
+    ./build_${BUILD_OPT}/bin/llama-cli -t 1 -m "$MODEL" -s "$SEED" -n "$NUMGEN" -e -p "$MODEL_PROMPT" # running with 1 thread
+fi # dorun
